[MODI-232] REFACTOR: 파티 목록 조회 totalSize 오류 수정
- 이미 시작됐지만 상태가 변경되지 않은 파티 데이터
- 이미 모집이 완료된 파티
위 두 경우 때문에 파티 목록 조회 API에서 잘못된 totalSize 응답해 수정

+++ src/main/java/com/prgrms/modi/party/dto/response/PartyListResponse.java
@@ -13,12 +13,12 @@ public class PartyListResponse {
     private String ottName;
 
     @ApiModelProperty(value = "OTT 파티목록 전체 size")
-    private Integer totalSize;
+    private long totalSize;
 
     @ApiModelProperty(value = "간략한 파티 정보")
     private List<PartyBriefResponse> partyList;
 
-    public PartyListResponse(Long ottId, String ottName, Integer totalSize, List<PartyBriefResponse> partyList) {
+    public PartyListResponse(Long ottId, String ottName, long totalSize, List<PartyBriefResponse> partyList) {
         this.ottId = ottId;
         this.ottName = ottName;
         this.totalSize = totalSize;
@@ -33,7 +33,7 @@ public class PartyListResponse {
         return ottName;
     }
 
-    public Integer getTotalSize() {
+    public long getTotalSize() {
         return totalSize;
     }
 
@@ -41,7 +41,7 @@ public class PartyListResponse {
         return partyList;
     }
 
-    public static PartyListResponse from(OTT ott, Integer totalSize, List<PartyBriefResponse> parties) {
+    public static PartyListResponse from(OTT ott, long totalSize, List<PartyBriefResponse> parties) {
         return new PartyListResponse(
             ott.getId(),
             ott.getName(),
